Create proxy_task module
