+++ src/lib/proxy_task.rs
@@ -1,0 +1,57 @@
+use std::env;
+
+use crate::{logger, profile, types::Task};
+
+pub(crate) fn create_proxy_task(
+    task: &str,
+    allow_private: bool,
+    skip_init_end_tasks: bool,
+) -> Task {
+    //get log level name
+    let log_level = logger::get_log_level();
+
+    let mut log_level_arg = "--loglevel=".to_string();
+    log_level_arg.push_str(&log_level);
+
+    //get profile
+    let profile_name = profile::get();
+
+    let mut profile_arg = "--profile=".to_string();
+    profile_arg.push_str(&profile_name);
+
+    //setup common args
+    let mut args = vec![
+        "make".to_string(),
+        "--disable-check-for-updates".to_string(),
+        "--no-on-error".to_string(),
+        log_level_arg.to_string(),
+        profile_arg.to_string(),
+    ];
+
+    if allow_private {
+        args.push("--allow-private".to_string());
+    }
+
+    if skip_init_end_tasks {
+        args.push("--skip-init-end-tasks".to_string());
+    }
+
+    //get makefile location
+    match env::var("CARGO_MAKE_MAKEFILE_PATH") {
+        Ok(makefile_path) => {
+            if makefile_path.len() > 0 {
+                args.push("--makefile".to_string());
+                args.push(makefile_path);
+            }
+        }
+        _ => (),
+    };
+
+    args.push(task.to_string());
+
+    let mut proxy_task = Task::new();
+    proxy_task.command = Some("cargo".to_string());
+    proxy_task.args = Some(args);
+
+    proxy_task.get_normalized_task()
+}
