+++ src/lib/runner.rs
@@ -20,6 +20,7 @@ use crate::functions;
 use crate::installer;
 use crate::logger;
 use crate::profile;
+use crate::proxy_task::create_proxy_task;
 use crate::scriptengine;
 use crate::time_summary;
 use crate::types::{
@@ -27,7 +28,6 @@ use crate::types::{
     RunTaskInfo, RunTaskName, RunTaskRoutingInfo, Step, Task, TaskWatchOptions,
 };
 use indexmap::IndexMap;
-use std::env;
 use std::thread;
 use std::time::SystemTime;
 
@@ -494,60 +494,6 @@ fn create_watch_task(task: &str, options: Option<TaskWatchOptions>) -> Task {
     task_config.args = Some(watch_args);
 
     task_config
-}
-
-pub(crate) fn create_proxy_task(
-    task: &str,
-    allow_private: bool,
-    skip_init_end_tasks: bool,
-) -> Task {
-    //get log level name
-    let log_level = logger::get_log_level();
-
-    let mut log_level_arg = "--loglevel=".to_string();
-    log_level_arg.push_str(&log_level);
-
-    //get profile
-    let profile_name = profile::get();
-
-    let mut profile_arg = "--profile=".to_string();
-    profile_arg.push_str(&profile_name);
-
-    //setup common args
-    let mut args = vec![
-        "make".to_string(),
-        "--disable-check-for-updates".to_string(),
-        "--no-on-error".to_string(),
-        log_level_arg.to_string(),
-        profile_arg.to_string(),
-    ];
-
-    if allow_private {
-        args.push("--allow-private".to_string());
-    }
-
-    if skip_init_end_tasks {
-        args.push("--skip-init-end-tasks".to_string());
-    }
-
-    //get makefile location
-    match env::var("CARGO_MAKE_MAKEFILE_PATH") {
-        Ok(makefile_path) => {
-            if makefile_path.len() > 0 {
-                args.push("--makefile".to_string());
-                args.push(makefile_path);
-            }
-        }
-        _ => (),
-    };
-
-    args.push(task.to_string());
-
-    let mut proxy_task = Task::new();
-    proxy_task.command = Some("cargo".to_string());
-    proxy_task.args = Some(args);
-
-    proxy_task.get_normalized_task()
 }
 
 fn run_flow(flow_info: &FlowInfo, flow_state: &mut FlowState, sub_flow: bool) {
